pyproject.toml
[project]
name = "src"
version = "0.1.0"
description =""
author = [
    {name = "An Dong", email = "an.dong90@yahoo.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "langchain-core (>=0.3.40,<0.4.0)",
    "langchain-openai (>=0.3.7,<0.4.0)",
    "langchain-postgres (>=0.0.13,<0.0.14)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi-cache2[redis] (==0.1.9)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "sqlmodel (>=0.0.23,<0.0.24)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "langgraph (>=0.3.11,<0.4.0)",
    "boto3 (>=1.37.13,<2.0.0)",
    "langgraph-cli (>=0.1.77,<0.2.0)",
    "langchain (>=0.3.21,<0.4.0)",
    "pgvector (>=0.3.0,<0.4.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "qdrant-client (>=1.13.3,<2.0.0)",
    "cohere (>=5.14.0,<6.0.0)",
    "langchain-community (>=0.3.20,<0.4.0)",
    "rank-bm25 (>=0.2.2,<0.3.0)",
    "fastembed (>=0.6.0,<0.7.0)",
    "langchain-qdrant (>=0.2.0,<0.3.0)",
    "psycopg-pool (>=3.2.6,<4.0.0)",
    "langgraph-checkpoint-postgres (>=2.0.16,<3.0.0)",
    "psycopg-binary (>=3.2.6,<4.0.0)",
    "langgraph-checkpoint-sqlite (>=2.0.6,<3.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
pytest-cov = "^4.1.0"
pytest-xdist = "^3.5.0"

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
addopts = "-s --cov=src --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "asyncio: mark test as async",
    "memory: mark test as memory-related",
    "integration: mark test as integration test",
    "unit: mark test as unit test"
]
